adjustments for snowfkareyservice refacory
