+++ src/treecipe/src/RecipeService/RecipeService.ts
@@ -9,41 +9,19 @@ export class RecipeService {
 
 
     private fakerService: IFakerService;
-
     constructor(private fakerServiceImplementation) {
         this.fakerService = fakerServiceImplementation;
     }
 
-    openingRecipeSyntax:string = "${{";
-    closingRecipeSyntax:string = "}}";
     
     generateTabs(tabCount: number):string {
         const spacesPerTab = 4;
         return ' '.repeat(spacesPerTab * tabCount);
     }
 
-    static salesforceFieldToSnowfakeryMap: Record<string, string> = {
-        'text': '${{fake.text(max_nb_chars=50)}}',
-        'textarea': '${{fake.paragraph()}}',
-        'longtextarea': '${{fake.text(max_nb_chars=1000)}}',
-        'richtextarea': '${{fake.text(max_nb_chars=1000)}}',
-        'email': '${{fake.email()}}',
-        'phone': '${{fake.phone_number()}}',
-        'url': '${{fake.url()}}',
-        'number': '${{fake.random_int(min=0, max=999999)}}',
-        'currency': '${{fake.pydecimal(left_digits=6, right_digits=2, positive=True)}}',
-        'percent': '${{fake.pydecimal(left_digits=2, right_digits=2, positive=True)}}',
-        'date': '${{date(fake.date_between(start_date="-1y", end_date="today"))}}',
-        'datetime': '${{fake.date_time_between(start_date="-1y", end_date="now")}}',
-        'time': '${{fake.time()}}',
-        'picklist': 'GENERATED BY FIELD XML MARKUP',
-        'multiselectpicklist': 'GENERATED BY FIELD XML MARKUP',
-        'checkbox': '${{fake.boolean()}}',
-        'lookup': '"TODO -- REFERENCE ID REQUIRED"',
-        'masterdetail': '"TODO -- REFERENCE ID REQUIRED"',
-        'formula': 'Formula fields are calculated, not generated',
-        'location': '"SEE ONE PAGER - https://gist.github.com/jdschleicher/4abfd188a933598833285ee76e560445"'
-    };
+    getSalesforceFieldToSnowfakeryMap() {
+        this.fakerService.getMapSalesforceFieldToFakerValue();
+    }
 
     getRecipeFakeValueByXMLFieldDetail(xmlFieldDetail: XMLFieldDetail): string {
         
@@ -54,17 +32,25 @@ export class RecipeService {
             case 'picklist':
                 
                 if (xmlFieldDetail.controllingField) {
-                    fakeRecipeValue = this.buildDependentPicklistRecipeFakerValue(xmlFieldDetail);
+                    fakeRecipeValue = this.getDependentPicklistRecipeFakerValue(xmlFieldDetail);
                 } else {
-                    fakeRecipeValue = this.buildPicklistRecipeValueByXMLFieldDetail(xmlFieldDetail);
+                    if ( !(xmlFieldDetail.picklistValues) ) {
+                        return '';
+                    }
+                    const availablePicklistChoices = xmlFieldDetail.picklistValues.map(detail => detail.fullName);
+                    fakeRecipeValue = this.fakerService.buildPicklistRecipeValueByXMLFieldDetail(availablePicklistChoices);
                 }
 
                 return fakeRecipeValue;
                 
             case 'multiselectpicklist':
-                
-                fakeRecipeValue = this.buildMultiSelectPicklistRecipeValueByXMLFieldDetail(xmlFieldDetail);
-            
+
+                if ( !(xmlFieldDetail.picklistValues) ) {
+                    return '';
+                }
+                const availablePicklistChoices = xmlFieldDetail.picklistValues.map(detail => detail.fullName);
+                fakeRecipeValue = this.fakerService.buildMultiSelectPicklistRecipeValueByXMLFieldDetail(availablePicklistChoices);
+        
                 return fakeRecipeValue;
                 
             // case 'masterdetail':
@@ -94,24 +80,9 @@ export class RecipeService {
 
     }
 
-    buildMultiSelectPicklistRecipeValueByXMLFieldDetail(xmlFieldDetail: XMLFieldDetail): string {
-
-        if ( !(xmlFieldDetail.picklistValues) ) {
-            return '';
-        }
-        const availablePicklistChoices = xmlFieldDetail.picklistValues.map(detail => detail.fullName);
-
-        // const commaJoinedPicklistChoices = availablePicklistChoices.join("','");
-        // const fakeRecipeValue = `${this.openingRecipeSyntax} (';').join((fake.random_sample(elements=('${commaJoinedPicklistChoices}')))) ${this.closingRecipeSyntax}`;
-        const fakeRecipeValue = this.fakerService.getFakeMultiSelectPicklistRecipeValueByXMLFieldDetail(availablePicklistChoices);
-        return fakeRecipeValue;
-
-    }
-
-    buildDependentPicklistRecipeFakerValue(xmlFieldDetail: XMLFieldDetail): string {
+    getDependentPicklistRecipeFakerValue(xmlFieldDetail: XMLFieldDetail): string {
     
         const controllingField = xmlFieldDetail.controllingField;
-        let fakeRecipeValue:string;
         let controllingValueToPicklistOptions:Record<string, string[]> = {};
 
         if ( !(xmlFieldDetail.picklistValues) ) {
@@ -134,60 +105,10 @@ export class RecipeService {
 
         });
 
-        let allMultiSelectChoiceRecipe:string;
-        for ( const controllingValueKey in controllingValueToPicklistOptions ) {
-            
-            let picklistValuesAvailableForChoice = controllingValueToPicklistOptions[controllingValueKey];
-            
-            let randomChoicesBreakdown:string;
-
-            picklistValuesAvailableForChoice.forEach( value => {
-                if (randomChoicesBreakdown) {
-                    randomChoicesBreakdown += `\n${this.generateTabs(5)}- ${value}`;
-                } else {
-                    randomChoicesBreakdown = `- ${value}`;
-                }
-            });
-
-
-            let multiSelectChoiceRecipe = 
-`${this.generateTabs(2)}- choice:
-${this.generateTabs(3)}when: ${this.openingRecipeSyntax} ${controllingField} == '${controllingValueKey}' }}
-${this.generateTabs(3)}pick:
-${this.generateTabs(4)}random_choice:
-${this.generateTabs(5)}${randomChoicesBreakdown}`;
-
-
-            if (!(allMultiSelectChoiceRecipe)) {
-                allMultiSelectChoiceRecipe = multiSelectChoiceRecipe;
-            } else {
-                const lineBreakMultiSelectChoiceRecipe = `\n${multiSelectChoiceRecipe}`;
-                allMultiSelectChoiceRecipe += lineBreakMultiSelectChoiceRecipe;
-            }
-        }
-
-        if (fakeRecipeValue) {
-            fakeRecipeValue += `\n${this.generateTabs(2)}${allMultiSelectChoiceRecipe}`;
-        } else {
-            fakeRecipeValue = `\n${this.generateTabs(1.5)}if:`;
-            fakeRecipeValue += `\n${allMultiSelectChoiceRecipe}`;
-        }
-
-        return fakeRecipeValue;
-
+        return this.fakerService.buildDependentPicklistRecipeFakerValue(controllingValueToPicklistOptions, controllingField);
+        
     }
 
-    buildPicklistRecipeValueByXMLFieldDetail(xmlFieldDetail: XMLFieldDetail): string {
-    
-        if ( !(xmlFieldDetail.picklistValues) ) {
-            return '';
-        }
-        const availablePicklistChoices = xmlFieldDetail.picklistValues.map(detail => detail.fullName);
-        const commaJoinedPicklistChoices = availablePicklistChoices.join("','");
-        const fakeRecipeValue = `${this.openingRecipeSyntax} random_choice('${commaJoinedPicklistChoices}') ${this.closingRecipeSyntax}`;
-        return fakeRecipeValue;
-
-    }
 
     initiateRecipeByObjectName(objectName: string): string {
